pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "citrature"
version = "0.1.0"
description = "AI-powered research paper analysis and citation graph platform"
authors = [{name = "Citrature Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "google-cloud-storage>=2.10.0",
    "httpx>=0.25.2",
    "pymupdf>=1.23.8",
    "grobid-client-python>=0.0.7",
    "openai>=1.3.7",
    "numpy>=1.24.4",
    "scikit-learn>=1.3.2",
    "whoosh>=2.7.4",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
